Added action on search form, renamed styles
