+++ src/pages/Home.page.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { PageContainer } from '../constants/Layout/Container.styles'
 import { Header } from '../components/Header/Header'
 import { StudentsNavigation } from '../components/Dashboard/Navigation/Navigation'
-import { SearchFilter } from '../components/Dashboard/SearchFilterForm/SearchFilterForm'
+import { SearchFilter } from '../components/Dashboard/SearchFilterBar/SearchFilterForm'
 
 export interface IHomePageProps {}
